requirements.txt
fastapi==0.115.12
openai==1.72.0
pip-chill==1.0.3
pypdf2==3.0.1
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.0
